refactor: 예외 의존성 분리
