+++ src/main/java/kitchenpos/domain/order/OrderStatusCheckerImpl.java
@@ -3,7 +3,7 @@ package kitchenpos.domain.order;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Set;
-import kitchenpos.application.exception.TableGroupAppException.UngroupingNotPossibleException;
+import kitchenpos.domain.exception.OrderException.InvalidOrderStatusChangeException;
 import kitchenpos.domain.table.OrderStatusChecker;
 import org.springframework.stereotype.Service;
 
@@ -20,18 +20,16 @@ public class OrderStatusCheckerImpl implements OrderStatusChecker {
     public void validateOrderStatusChangeable(final List<Long> orderIds) {
         if (orderRepository.existsByOrderTableIdInAndOrderStatusIn(orderIds,
             Arrays.asList(OrderStatus.COOKING, OrderStatus.MEAL))) {
-            throw new IllegalArgumentException();
+            throw new InvalidOrderStatusChangeException();
         }
     }
 
     @Override
-    public void validateOrderStatusChangeableByTableGroupId(final Long tableGroupId) {
+    public boolean checkUngroupableTableGroup(final Long tableGroupId) {
         final Set<Order> findOrders = orderRepository.findAllByTableGroupId(tableGroupId);
 
-        if (findOrders.stream()
+        return findOrders.stream()
             .anyMatch(orderTable -> orderTable.isSameStatus(OrderStatus.COOKING)
-                || orderTable.isSameStatus(OrderStatus.MEAL))) {
-            throw new UngroupingNotPossibleException();
-        }
+                || orderTable.isSameStatus(OrderStatus.MEAL));
     }
 }
