+++ src/main/java/kitchenpos/domain/table/OrderStatusChecker.java
@@ -6,5 +6,5 @@ public interface OrderStatusChecker {
 
     void validateOrderStatusChangeable(final List<Long> orderIds);
 
-    void validateOrderStatusChangeableByTableGroupId(final Long tableGroupId);
+    boolean checkUngroupableTableGroup(final Long tableGroupId);
 }
