+++ src/main/java/kitchenpos/domain/table/OrderTableChangeServiceImpl.java
@@ -1,6 +1,6 @@
 package kitchenpos.domain.table;
 
-import kitchenpos.application.exception.OrderTableAppException.NotFoundOrderTableException;
+import kitchenpos.domain.exception.OrderTableException.NotFoundOrderTableException;
 import kitchenpos.domain.order.OrderTableChangeService;
 import org.springframework.stereotype.Service;
