+++ src/test/java/kitchenpos/application/TableGroupServiceTest.java
@@ -8,8 +8,8 @@ import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
 import kitchenpos.application.dto.TableGroupCreateDto;
-import kitchenpos.application.exception.TableGroupAppException.UngroupingNotPossibleException;
 import kitchenpos.domain.exception.TableGroupException.InvalidOrderTablesException;
+import kitchenpos.domain.exception.TableGroupException.UngroupingNotPossibleException;
 import kitchenpos.domain.menu.Menu;
 import kitchenpos.domain.menu.MenuGroup;
 import kitchenpos.domain.menu.MenuGroupRepository;
